pyproject.toml
[project]
name = "house-price-prediction"
version = "0.1.0"
description = "House Price Prediction"
authors = [
    { name = "Sriram Peddapanta", email = "SriramPeddapanta@users.noreply.github.com" }
]
dependencies = [
    "numpy>=2.2.1",
    "pandas>=2.2.3",
    "scikit-learn>=1.6.0",
    "matplotlib>=3.10.0",
    "azureml-core>=1.59.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.rye.scripts]
train = { cmd = "python src/house_price_prediction/train.py", env-file = ".env" }

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/house_price_prediction"]
